refactor: create setupMcpCli abstraction and update imports
Create CLI abstraction with setupMcpCli function:
- Add cli.ts with setupMcpCli(), ServerFactory, RequestHandler types
- Add handleDryRun, handleHelpRequest, handleVersionRequest helpers
- Add MCPHttpServerWrapper for HTTP transport
- Update CLI package dependencies
- Add path helpers (getLocalDataPath, getLogPath, getExportsPath) to CLI package
- Update all imports in mongodb-mcp-server to use @mongodb-js/mcp-cli directly
- Remove wrapper files in mongodb-mcp-server (use CLI package directly)
- Update index.ts to use setupMcpCli with serverFactory and requestHandler
