+++ packages/cli/src/index.ts
@@ -1,8 +1,22 @@
-// Transport constants and types
+// Main CLI setup function
 export {
-    TRANSPORT_PAYLOAD_LIMITS,
-    type TransportType,
-} from "./transports/constants.js";
+    setupMcpCli,
+    type CLIOptions,
+    type ServerFactory,
+    type RequestHandler,
+    handleHelpRequest,
+    handleVersionRequest,
+    createDefaultLoggers,
+    handleDryRun,
+} from "./cli.js";
+
+// Config parsing and types
+export { parseUserConfig, type UserConfig } from "./config/parseUserConfig.js";
+export {
+    UserConfigSchema,
+    configRegistry,
+    ALL_CONFIG_KEYS,
+} from "./config/userConfig.js";
 
 // Config utilities
 export {
@@ -16,3 +30,14 @@ export {
     type OverrideBehavior,
     type ConfigFieldMeta,
 } from "./config/configUtils.js";
+
+// Transport constants
+export { TRANSPORT_PAYLOAD_LIMITS, type TransportType } from "./transports/constants.js";
+
+// Dry run runner
+export {
+    DryRunModeRunner,
+    type DryRunServer,
+    type DryRunLogger,
+    type DryRunModeRunnerOptions,
+} from "./transports/dryModeRunner.js";
