+++ packages/mongodb-mcp-server/src/common/config/configUtils.ts
@@ -1,19 +1,6 @@
 import path from "path";
 import os from "os";
 
-// Re-export generic utilities from mcp-cli
-export {
-    commaSeparatedToArray,
-    parseBoolean,
-    oneWayOverride,
-    onlyLowerThanBaseValueOverride,
-    onlyStricterLogLevelOverride,
-    onlySubsetOfBaseValueOverride,
-    type CustomOverrideLogic,
-    type OverrideBehavior,
-    type ConfigFieldMeta,
-} from "@mongodb-js/mcp-cli";
-
 export function getLocalDataPath(): string {
     return process.platform === "win32"
         ? path.join(process.env.LOCALAPPDATA || process.env.APPDATA || os.homedir(), "mongodb")
