+++ packages/mongodb-mcp-server/src/common/session.ts
@@ -17,7 +17,7 @@ import type { ExportsManager } from "@mongodb-js/mcp-tools-mongodb";
 import type { Client } from "@mongodb-js/atlas-local";
 import type { Keychain } from "@mongodb-js/mcp-core";
 import { generateConnectionInfoFromCliArgs } from "@mongosh/arg-parser";
-import { type UserConfig } from "../common/config/userConfig.js";
+import type { UserConfig } from "@mongodb-js/mcp-cli";
 import { type ConnectionErrorHandler } from "./connectionErrorHandler.js";
 
 export interface SessionOptions<TUserConfig extends UserConfig = UserConfig> {
