+++ packages/mongodb-mcp-server/src/resources/common/config.ts
@@ -1,5 +1,5 @@
 import { ReactiveResource } from "../resource.js";
-import type { UserConfig } from "../../common/config/userConfig.js";
+import type { UserConfig } from "@mongodb-js/mcp-cli";
 import type { AtlasTelemetry } from "@mongodb-js/mcp-atlas-telemetry";
 import type { Session } from "../../lib.js";
 import { generateConnectionInfoFromCliArgs } from "@mongosh/arg-parser";
