+++ packages/mongodb-mcp-server/src/resources/resource.ts
@@ -1,6 +1,6 @@
 import type { Server } from "../server.js";
 import type { Session } from "../common/session.js";
-import type { UserConfig } from "../common/config/userConfig.js";
+import type { UserConfig } from "@mongodb-js/mcp-cli";
 import type { AtlasTelemetry } from "@mongodb-js/mcp-atlas-telemetry";
 import type { SessionEvents } from "../common/session.js";
 import { LogId } from "@mongodb-js/mcp-core";
