+++ packages/mongodb-mcp-server/src/setup/setupMcpServer.ts
@@ -11,7 +11,7 @@ import type { Platform } from "./setupAiToolsUtils.js";
 import { formatError, getPlatform } from "./setupAiToolsUtils.js";
 import { packageInfo } from "../common/packageInfo.js";
 import { getAuthType } from "@mongodb-js/mcp-tools-mongodb";
-import { type UserConfig } from "../common/config/userConfig.js";
+import type { UserConfig } from "@mongodb-js/mcp-cli";
 import { createAtlasLocalClient } from "@mongodb-js/mcp-tools-atlas-local";
 import { NoopLogger } from "@mongodb-js/mcp-core";
 import type { TelemetryResult } from "@mongodb-js/mcp-atlas-telemetry";
