+++ packages/mongodb-mcp-server/src/test-helpers/index.ts
@@ -1,4 +1,4 @@
-import { UserConfigSchema, type UserConfig } from "../common/config/userConfig.js";
+import { UserConfigSchema, type UserConfig } from "@mongodb-js/mcp-cli";
 import { packageInfo } from "../common/packageInfo.js";
 
 export const defaultTestConfig: UserConfig = {
