+++ packages/mongodb-mcp-server/tests/unit/toolBase.test.ts
@@ -4,7 +4,7 @@ import type { ZodRawShape } from "zod";
 import type { ToolConstructorParams } from "../../src/tools/tool.js";
 import type { ToolAnnotations } from "@modelcontextprotocol/sdk/types.js";
 import type { Session } from "../../src/common/session.js";
-import type { UserConfig } from "../../src/common/config/userConfig.js";
+import type { UserConfig } from "@mongodb-js/mcp-cli";
 import type { AtlasTelemetry } from "@mongodb-js/mcp-atlas-telemetry";
 import type { Elicitation } from "@mongodb-js/mcp-core";
 import type { CompositeLogger } from "@mongodb-js/mcp-core";
